# docs/openapi.yaml
openapi: 3.1.3
info:
  title: Where We Are API
  description: |
    "Questa API fornisce funzionalità per la gestione di punti di interesse (POI) e la raccolta di statistiche sugli accessi e l'interazione degli utenti. Gli endpoint consentono di aggiungere, modificare e rimuovere POI, oltre a ottenere dati analitici aggregati per comprendere meglio il comportamento degli utenti.
      
      Qui sono riportate sia le API già sviluppate sia quelle pianificate, leggere pertanto attentamente la documentazione."
    
  license:
    name: GPL 3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  version: 0.0.1
...
paths:
...
  /api/stat/poi/{poiId}:
    get:
      tags: [poi-stats]
      summary: Get statistics for a specific POI
      operationId: getPoiStats
      parameters:
        - name: poiId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: POI statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoiStats'
        '404':
          description: POI not found

  # POI Analytics
...
  /api/stat/compare:
    get:
      tags: [poi-stats]
      summary: Compare POI statistics
      operationId: comparePoiStats
      parameters:
        - name: poiIds
          in: query
          required: true
          schema:
            type: array
            items:
              type: string
          style: form
          explode: false
      responses:
        '200':
          description: Comparison of POI statistics
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    poiId: { type: string }
                    visits: { type: integer }
                    upvotes: { type: integer }
                    downvotes: { type: integer }

  # POI Interactions
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PoiStats:
      type: object
      properties:
        _id:
          type: string
          readOnly: true
        poiId: 
          type: string
          description: Reference to POI ObjectId
        visits:
          type: array
          items:
            type: object
            properties:
              timestamp: 
                type: string
                format: date-time
                description: Visit timestamp, defaults to current time
        rating:
          type: object
          properties:
            upvotes: 
              type: integer 
              minimum: 0
              default: 0
            downvotes: 
              type: integer 
              minimum: 0
              default: 0
        ratingMensile:
          type: array
          items:
            type: object
            properties:
              year: { type: integer }
              month: { type: integer, minimum: 1, maximum: 12 }
              upvotes: { type: integer, minimum: 0, default: 0 }
              downvotes: { type: integer, minimum: 0, default: 0 }
...